requirements.txt
openai>=1.12.0
google-genai>=0.14.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1